requirements.txt
langchain>=0.1.20
langchain-openai
langchain-community==0.0.38
langchain-chroma>=0.1.0
sentence-transformers>=2.2.2
pypdf>=3.17.0
openai-whisper>=20231117
python-dotenv>=1.0.0
gradio>=4.0.0
tiktoken